Queue write-operations (still synchronous)
Though the execution is still synchronous, a tasks queue is already
implemented. When a write-operation is requested, a new task is added
to the queue; then it is immediately retrieved and executed. Its result
continues to be stored in a list of task-results (that can already be
queried asynchronously using the query-operation 'verify').