Change address of the server's listening socket
We used to set the s_addr to the wildcard INADDR_ANY (all interfaces).
However, by doing so, when we tried to retrive the host's IP address
using function 'getsockname', it would always return "0.0.0.0".
For this reason we needed to set a specific IP address when binding
the server's listening socket.

(Th referred function 'getsockname' will be used in the next commit.)